vsock: flow control fix
This patch has been cherry-picked from the Firecracker tree. The
reference commit is 1cc8b8a678eb28b20f5843556bdb7fbb2dfa6284.

Fixed a logical error in the vsock flow control, that would cause credit
update packets to not be sent at the right time.

+++ vm-virtio/src/vsock/csm/txbuf.rs
@@ -25,7 +25,7 @@ pub struct TxBuf {
 impl TxBuf {
     /// Total buffer size, in bytes.
     ///
-    const SIZE: usize = defs::CONN_TX_BUF_SIZE;
+    const SIZE: usize = defs::CONN_TX_BUF_SIZE as usize;
 
     /// Ring-buffer constructor.
     ///
